+++ CHANGELOG.md
@@ -3,12 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.1.0] - 2024-10-15
 
 ### Changed
 
 - General settings now applied by home manager
 - Implement modes
+- Use default theme
+- Further minimize graphical interface
 	
 ## [2.0.0] - 2024-09-19